rename printer-manager to follow current format
