+++ client_src/js/printer-management.runner.js
@@ -10,7 +10,7 @@ import {
   bulkOctoPrintControlCommand,
   bulkOctoPrintGcodeCommand,
   bulkOctoPrintPluginAction
-} from "./printerManager/functions/bulk-commands-functions";
+} from "./printer-manager/functions/bulk-commands-functions";
 import {
   addBlankPrinterToTable,
   bulkDeletePrinters,
@@ -22,9 +22,9 @@ import {
   saveAllOnAddPrinterTable,
   scanNetworkForDevices,
   workerEventFunction
-} from "./printerManager/functions/printer-manager.functions";
-import { loadCustomGcodeScriptsModel } from "./printerManager/functions/custom-gcode-scripts.functions";
-import { setupSortablePrintersTable } from "./printerManager/functions/sortable-table";
+} from "./printer-manager/functions/printer-manager.functions";
+import { loadCustomGcodeScriptsModel } from "./printer-manager/functions/custom-gcode-scripts.functions";
+import { setupSortablePrintersTable } from "./printer-manager/functions/sortable-table";
 
 const workerURL = "/printersInfo/get/";
