+++ client_src/js/printer-manager/connection-log.js
