+++ client_src/js/printer-manager/functions/bulk-commands-functions.js
