+++ client_src/js/printer-manager/functions/plugin-search.function.js
