+++ client_src/js/printer-manager/functions/printer-manager.functions.js
