+++ client_src/js/printer-manager/templates/gcode-scripts-table.template.js
