+++ client_src/js/printer-manager/templates/octoprint-plugin-list.template.js
