+++ client_src/js/printer-manager/templates/printer-add-instructions.template.js
