+++ client_src/js/printer-manager/templates/printer-table-row.templates.js
